changed data types of particles

+++ serial/init_particles.h
@@ -3,9 +3,9 @@
 
 typedef struct {
     double x, y;
-    long vx, vy;
-    long long m;
-    long gravity_x, gravity_y;
+    double vx, vy;
+    double m;
+    double gravity_x, gravity_y;
 } particle_t;
 
 // Function prototype
